Share stable workload lab planner
